Search and Destroy
It's like everything is almost working for Search and Destroy, but I'm
having some issues. All except one target is supposed to be hidden, but
I had issues with putting the gameObjects into an array and using them
from the array. I'm also having issues with having the target script see
the search script to add the points when destroyed. Also my version is
complaining about the IngameMenu, but I didn't mess with it (besides
commenting it to test my code).